pyproject.toml
[tool.poetry]
name = "datalake-format-explorer"
version = "0.1.0"
description = ""
authors = ["Luciano Filho <lvgalvaofilho@gmail.com>"]
readme = "README.md"
packages = [{ include = "datalake_format_explorer" }]

[tool.poetry.dependencies]
python = "^3.11"
pre-commit = "^3.4.0"
makefile = "^1.1.0"
isort = "^5.12.0"
black = "^23.9.1"
mkdocs = "^1.5.3"
pip-audit = "^2.6.1"
pymdown-extensions = "^10.3"
mkdocs-material = "^9.4.1"
pygments = "^2.16.1"
mkdocstrings-python = "^1.7.0"
python-dotenv = "^1.0.0"
pydantic = "^2.3.0"
httpx = "^0.25.0"
polars = "^0.19.3"
boto3 = "^1.28.53"

[tool.isort]
profile = "black"
known_third_party = []


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
